Move UnresolvedIdentifierScanner to parser internal
