+++ core/src/main/java/com/nikodoko/javaimports/parser/internal/UnresolvedIdentifierScanner.java
@@ -1,8 +1,8 @@
-package com.nikodoko.javaimports.parser;
+package com.nikodoko.javaimports.parser.internal;
 
-import com.nikodoko.javaimports.parser.internal.ClassEntity;
-import com.nikodoko.javaimports.parser.internal.ClassSelectors;
-import com.nikodoko.javaimports.parser.internal.Scope;
+import com.nikodoko.javaimports.parser.ClassExtender;
+import com.nikodoko.javaimports.parser.ClassHierarchies;
+import com.nikodoko.javaimports.parser.ClassHierarchy;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.BiFunction;
