+++ core/src/test/java/com/nikodoko/javaimports/parser/internal/UnresolvedIdentifierScannerTest.java
@@ -1,4 +1,4 @@
-package com.nikodoko.javaimports.parser;
+package com.nikodoko.javaimports.parser.internal;
 
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.fail;
@@ -6,7 +6,8 @@ import static org.junit.Assert.fail;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Sets;
 import com.nikodoko.javaimports.ImporterException;
-import com.nikodoko.javaimports.parser.internal.Scope;
+import com.nikodoko.javaimports.parser.ClassExtender;
+import com.nikodoko.javaimports.parser.Parser;
 import java.util.Collection;
 import java.util.Set;
 import org.junit.Test;
